app/requirements.txt
editdistance==0.6.2
func_timeout==4.3.5
nltk==3.7
numpy==1.22.3
rapidfuzz==2.0.11
scikit_learn==1.2.1
spacy==2.2.3
sql_metadata==2.6.0
sqlparse==0.4.2
tokenizers==0.11.6
tqdm==4.63.0
transformers==4.28.1
tensorboard==2.8.0
sentencepiece==0.1.99
six==1.16.0
accelerate==0.25.0
streamlit-ace==0.1.1
eralchemy
SQLAlchemy==1.0.13
opencv-python-headless==4.8.1.78
protobuf==3.20.3
